requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
altgraph==0.17.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
frozenlist==1.3.0
future==0.18.2
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
heroku==0.1.4
heroku3==5.1.4
idna==3.3
Jinja2==3.0.3
lottie==0.6.11+src
MarkupSafe==2.0.1
meval==2.5
multidict==6.0.2
pefile==2021.9.3
Pillow==9.0.1
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyowm==3.3.0
Pyrogram==1.4.0
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
pythondialog==3.5.3
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.1
rsa==4.8
six==1.16.0
smmap==5.0.0
speedtest-cli==2.1.3
sqlparse==0.4.2
Telethon-Mod==1.24.3
TgCrypto==1.2.3
tinycss2==1.1.1
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
whitenoise==5.3.0
yarl==1.7.2